Split the poller interfaces
- Linefeed
- Documentation and method annotation
