+++ Poller/Poller/OAuth/OAuthAuthenticationType.cs
