+++ Poller/Poller/Poller/IPollerCreateOrUpdateObjects.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Poller.Poller
+{
+    public interface IPollerCreateOrUpdateObjects : IPoller
+    {
+        Task CreateOrUpdateObjectsAsync(PollerContext contex, CancellationToken token);
+    }
+}
