+++ Poller/Poller/Poller/IPollerDataSyncback.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Poller.Poller
+{
+    public interface IPollerDataSyncback : IPoller
+    {
+        Task DataSyncbackAsync(PollerContext contex, CancellationToken token);
+    }
+}
