+++ Poller/Poller/Poller/IPollerRefreshAdvertisementData.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Poller.Poller
+{
+    public interface IPollerRefreshAdvertisementData : IPoller
+    {
+        Task RefreshAdvertisementDataAsync(PollerContext contex, CancellationToken token);
+    }
+}
